# workspaces/backend/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage notebooks in a Kubernetes cluster.\nFor more information, visit https://www.kubeflow.org/docs/components/notebooks/",
        "title": "Kubeflow Notebooks API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "Provides a healthcheck response indicating the status of key services.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "healthcheck"
                ],
                "summary": "Returns the health status of the application",
                "responses": {
                    "200": {
                        "description": "Successful healthcheck response",
                        "schema": {
                            "$ref": "#/definitions/health_check.HealthCheck"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/workspaces/{namespace}/{workspaceName}/operations/pause": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Pauses a workspace, stopping all associated pods.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspaces"
                ],
                "summary": "Pause workspace",
                "parameters": [
                    {
                        "type": "string",
                        "example": "default",
                        "description": "Namespace of the workspace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "my-workspace",
                        "description": "Name of the workspace",
                        "name": "workspaceName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns an empty JSON object.",
                        "schema": {
                            "$ref": "#/definitions/api.EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request. Invalid workspace kind name format.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to access the workspace.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found. Workspace does not exist.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.EmptyResponse": {
            "type": "object"
        },
        "api.ErrorCause": {
            "type": "object",
            "properties": {
                "validation_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ValidationError"
                    }
                }
            }
        },
        "api.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.HTTPError"
                }
            }
        },
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "cause": {
                    "$ref": "#/definitions/api.ErrorCause"
                },
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/field.ErrorType"
                }
            }
        },
        "field.ErrorType": {
            "type": "string",
            "enum": [
                "FieldValueNotFound",
                "FieldValueRequired",
                "FieldValueDuplicate",
                "FieldValueInvalid",
                "FieldValueNotSupported",
                "FieldValueForbidden",
                "FieldValueTooLong",
                "FieldValueTooMany",
                "InternalError",
                "FieldValueTypeInvalid"
            ],
            "x-enum-varnames": [
                "ErrorTypeNotFound",
                "ErrorTypeRequired",
                "ErrorTypeDuplicate",
                "ErrorTypeInvalid",
                "ErrorTypeNotSupported",
                "ErrorTypeForbidden",
                "ErrorTypeTooLong",
                "ErrorTypeTooMany",
                "ErrorTypeInternal",
                "ErrorTypeTypeInvalid"
            ]
        },
        "health_check.HealthCheck": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/health_check.ServiceStatus"
                },
                "systemInfo": {
                    "$ref": "#/definitions/health_check.SystemInfo"
                }
            }
        },
        "health_check.ServiceStatus": {
            "type": "string",
            "enum": [
                "Healthy",
                "Unhealthy"
            ],
            "x-enum-varnames": [
                "ServiceStatusHealthy",
                "ServiceStatusUnhealthy"
            ]
        },
        "health_check.SystemInfo": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        }
    }
}